requirements.txt
flask
flask-cors
ultralytics
torch
Pillow
gunicorn
